+++ app/src/main/java/com/future/medan/backend/services/impl/PurchaseImpl.java
@@ -1,5 +1,6 @@
 package com.future.medan.backend.services.impl;
 
+import com.future.medan.backend.exceptions.ResourceNotFoundException;
 import com.future.medan.backend.models.entity.Purchase;
 import com.future.medan.backend.repositories.PurchaseRepository;
 import com.future.medan.backend.services.PurchaseService;
@@ -7,7 +8,6 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
-import java.util.Optional;
 
 @Service
 public class PurchaseImpl implements PurchaseService {
@@ -25,8 +25,8 @@ public class PurchaseImpl implements PurchaseService {
     }
 
     @Override
-    public Optional<Purchase> getById(String id){
-        return purchaseRepository.findById(id);
+    public Purchase getById(String id){
+        return purchaseRepository.findById(id).orElseThrow(() -> new ResourceNotFoundException("Purchase", "id", id));
     }
 
     @Override
@@ -36,6 +36,9 @@ public class PurchaseImpl implements PurchaseService {
 
     @Override
     public void deleteById(String id){
-        purchaseRepository.deleteById(id);
+        if (!purchaseRepository.existsById(id))
+            throw new ResourceNotFoundException("Purchase", "id", id);
+        else
+            purchaseRepository.deleteById(id);
     }
 }
